Minor tweaks
Remnants after a failed mixin/prototype augmentation approach with TypeScript + Phaser.

Notes to future self: taking advantage of type hinting, what phaser supplies, and code sharing is a bit difficult.
It feels like the best way to do things is as opposed to trying to force phaser, or typescript, to be something
they are not, to embrace both for what they are and make up my own sane patterns as I find them.